Drive Code V1.5
I got the code to stop doing the intake, I fixed a few other issues and added slew to the drivetrain to try to fix an issue with driving forwards and backwards really quickly. The color sensor still needs some changes but I should be able to fix that within the week.